pyproject.toml
[tool.poetry]
name = "inter-deliver"
version = "0.1.0"
description = ""
authors = ["varg <vardeathyandex.ru>"]
readme = "README.md"
packages = [{include = "inter_deliver"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
orjson = "^3.8.7"
aioredis = "^2.0.1"
python-dotenv = "^1.0.0"
itsdangerous = "^2.1.2"
aiomysql = "^0.1.1"
cryptography = "^39.0.2"
asyncmy = "^0.2.7"
pymysql = "^1.0.2"
alembic = "^1.10.2"
aio-pika = "^9.0.5"
sqlalchemy = "^2.0.7"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
uuid = "^1.30"
loguru = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
